Using lombok and mapstruct annotation

+++ src/main/java/com/dzungyb/learning_spring_boot/service/UserService.java
@@ -2,23 +2,26 @@ package com.dzungyb.learning_spring_boot.service;
 
 import com.dzungyb.learning_spring_boot.dto.request.UserCreationRequest;
 import com.dzungyb.learning_spring_boot.dto.request.UserUpdateRequest;
+import com.dzungyb.learning_spring_boot.dto.response.UserResponse;
+import com.dzungyb.learning_spring_boot.entity.User;
 import com.dzungyb.learning_spring_boot.exception.AppException;
 import com.dzungyb.learning_spring_boot.exception.ErrorCode;
 import com.dzungyb.learning_spring_boot.mapper.UserMapper;
-import com.dzungyb.learning_spring_boot.model.User;
 import com.dzungyb.learning_spring_boot.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class UserService {
-    @Autowired
     UserRepository userRepository;
-
-    @Autowired
-    private UserMapper userMapper;
+    UserMapper userMapper;
 
     public User createUser(UserCreationRequest request) {
         if (userRepository.existsByUserName(request.getUserName())) {
@@ -30,22 +33,39 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public List<User> getUsers() {
-        return userRepository.findAll();
+    public List<UserResponse> getUsers() {
+        List<User> users = userRepository.findAll();
+//        List<UserResponse> userResponses = new ArrayList<>();
+//
+//        for (User user : users) {
+//            userResponses.add(userMapper.toUserResponse(user));
+//        }
+
+        List<UserResponse> userResponses = users.stream()
+                .map(userMapper::toUserResponse)
+                .collect(Collectors.toList());
+
+        return userResponses;
     }
 
-    public User getUser(String userId) {
-        return userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
+    public UserResponse getUser(String userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+
+        return userMapper.toUserResponse(user);
     }
 
-    public User updateUser(String userId, UserUpdateRequest request) {
-        User user = getUser(userId);
+    public UserResponse updateUser(String userId, UserUpdateRequest request) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
         userMapper.updateUser(user, request);
 
-        return userRepository.save(user);
+        return userMapper.toUserResponse(userRepository.save(user));
     }
 
     public void deleteUser(String userId) {
+        if (userRepository.existsById(userId) == false) {
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.deleteById(userId);
     }
 }
@@ -53,4 +73,6 @@ public class UserService {
 /* Lưu ý:
 _ Ở tầng service khi trả về dữ liệu cho tầng controller thì nên trả về dữ liệu đã được xử lý, không nên trả về dữ liệu trực tiếp từ repository.
 => nên trả về dto khác vì ta có nhu cầu trả về object nhất định chứ không trả về toàn bộ entity.
+
+
  */
